agentic-ai/retail-demo/retail_agent/requirements.txt
google-adk
toolbox_core
a2a-sdk
a2a-sdk[http-server]
